Create Angle_Retrieval.R
This code takes the transformed points corresponding to the heads of the guppies and the points where the head joins the tail and converts them into angles in radians. It then finds both individual and pair sociality metrics and the distances between the heads of each pair of guppies. The actual transformation of the points is done beforehand in Excel.